+++ ui-design-lib/src/main/java/co/electriccoin/zcash/ui/design/theme/internal/Typography.kt
@@ -104,8 +104,8 @@ internal val SecondaryTypography =
         headlineSmall =
             TextStyle(
                 fontFamily = ArchivoFontFamily,
-                fontWeight = FontWeight.SemiBold,
-                fontSize = 18.sp,
+                fontWeight = FontWeight.Bold,
+                fontSize = 16.sp,
                 textAlign = TextAlign.Center
             ),
         bodyLarge =