+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/support/model/ConfigInfo.kt
@@ -3,6 +3,9 @@ package co.electriccoin.zcash.ui.screen.support.model
 import co.electriccoin.zcash.configuration.api.ConfigurationProvider
 import kotlinx.datetime.Instant
 
+// TODO [#1301]: Localize support text content
+// TODO [#1301]: https://github.com/Electric-Coin-Company/zashi-android/issues/1301
+
 data class ConfigInfo(val configurationUpdatedAt: Instant?) {
     fun toSupportString() =
         buildString {